refactor/#8 : TinkCryptoManager 책임 분리
