+++ app/src/main/java/com/kiero/core/security/CryptoManager.kt
@@ -1,0 +1,6 @@
+package com.kiero.core.security
+
+interface CryptoManager {
+    fun encrypt(plaintext: String): String
+    fun decrypt(ciphertext: String): String
+}
