+++ app/src/main/java/com/kiero/core/security/EncryptionException.kt
@@ -1,0 +1,4 @@
+package com.kiero.core.security
+
+class EncryptionException(message: String, cause: Throwable? = null) :
+    Exception(message, cause)
